feat(db,ext): Separate functions
