+++ README.md
@@ -1,6 +1,6 @@
 ## Goodnews
 
-This GO package is scrapping websites with good news (by default from `https://positivnews.ru/` and `https://ntdtv.ru/c/pozitivnye-novosti`), saving the ones from the last 14 days (in case they were not already saved) to the SQLite3 DB (by default `./data/news_items.db`) and sending them to the external service, specified in the `sendToExternalService` function (Telegram by default), sorted by the posted date. Once the item is sent, it is marked in the db as sent and won't be sent again.
+This GO package is scrapping websites with good news (by default from `https://positivnews.ru/` and `https://ntdtv.ru/c/pozitivnye-novosti`), saving the ones from the last 14 days (in case they were not already saved) to the SQLite3 DB (by default `./data/news_items.db`) and sending them to the external service, specified in the `external.SendToExternalService` function (Telegram by default), sorted by the posted date. Once the item is sent, it is marked in the db as sent and won't be sent again.
 
 ### Pre-req and installation
 
@@ -8,7 +8,7 @@ To run/build the project, you need GO 1.20+ and the packages mentioned in `go.mo
 
 #### To run the project:
 
-`mkdir data && API_KEY=<YOUR_TELEGRAM_BOT_API_KEY> CHAT_ID=<ID_OF_THE_CHAT_TO_SEND_NEWS_TO> go run goodnews.go`
+`mkdir data && API_KEY=<YOUR_TELEGRAM_BOT_API_KEY> CHAT_ID=<ID_OF_THE_CHAT_TO_SEND_NEWS_TO> go run main.go`
 
 #### To build and run the project:
 
@@ -24,9 +24,9 @@ mkdir data
 #### To build and run with Docker:
 
 ```
-docker build -t goodnews:1.0.0 .
+docker build -t goodnews:<VERSION> .
 docker volume create goodnews_data
-docker run --rm -v goodnews_data:/app/data -e API_KEY=<YOUR_TELEGRAM_BOT_API_KEY> -e CHAT_ID=<ID_OF_THE_CHAT_TO_SEND_NEWS_TO> goodnews:1.0.0
+docker run --rm -v goodnews_data:/app/data -e API_KEY=<YOUR_TELEGRAM_BOT_API_KEY> -e CHAT_ID=<ID_OF_THE_CHAT_TO_SEND_NEWS_TO> goodnews:<VERSION>
 ```
 
 If you already have a SQLite3 db file, and you'd like to use it with the newly created docker container, you can run the following command from the directory with the file to copy the file to the volume before you run the container (make sure that you have jq installed in your system):
@@ -42,4 +42,4 @@ Since version 1.1.0 you can use the following flags (both of them are turned off
 
 Usage example:
 
-`go run goodnews.go --dry-run --debug`
+`go run main.go --dry-run --debug`
