+++ external/external.go
@@ -1,0 +1,63 @@
+package external
+
+import (
+	"fmt"
+	"goodnews/scraping"
+	"io"
+	"log"
+	"math/rand"
+	"net/http"
+	"net/url"
+	"os"
+	"time"
+)
+
+func SendToExternalService(item scraping.NewsItem) error {
+	apiKey := os.Getenv("API_KEY")
+	chatId := os.Getenv("CHAT_ID")
+
+	messageEnding := pickRandomMessageEnding()
+	caption := url.QueryEscape(fmt.Sprintf("%s: %s\n\n%s\n\n%s\n\n%s", item.Category, item.Title, item.P1, item.Posted, messageEnding))
+
+	url := fmt.Sprintf("https://api.telegram.org/bot%s/sendPhoto", apiKey)
+
+	resp, err := http.Get(fmt.Sprintf("%s?chat_id=%s&photo=%s&caption=%s&parse_mode=Markdown", url, chatId, item.Image, caption))
+	if err != nil {
+		return err
+	}
+
+	if resp.StatusCode != http.StatusOK {
+		body, err := io.ReadAll(resp.Body)
+		if err != nil {
+			return err
+		}
+		log.Printf("Request: %s?chat_id=%s&photo=%s&caption=%s&parse_mode=Markdown", url, chatId, item.Image, caption)
+		log.Printf("Response body:\n%s", body)
+		return fmt.Errorf("sendToExternalService request for item id: %d. StatusCode: %d", item.Id, resp.StatusCode)
+	}
+	defer resp.Body.Close()
+
+	return nil
+}
+
+func pickRandomMessageEnding() string {
+	messageEndings := []string{
+		"[Подписывайся! У нас только хорошие новости!](t.me/nomoredoomscrolling)",
+		"[Жми сюда, если надоел Doom Scrolling](t.me/nomoredoomscrolling)",
+		"[Если понравилось, заходи. У нас есть ещё!](t.me/nomoredoomscrolling)",
+		"[Ждём тебя на нашем канале!](t.me/nomoredoomscrolling)",
+		"[Не упусти свою порцию позитива!](t.me/nomoredoomscrolling)",
+		"[Брось Doom Scrolling и присоединяйся!](t.me/nomoredoomscrolling)",
+		"[У нас всегда только светлая сторона новостей!](t.me/nomoredoomscrolling)",
+		"[Забудь о плохих новостях на нашем канале!](t.me/nomoredoomscrolling)",
+		"[Больше хороших новостей на нашем канале!](t.me/nomoredoomscrolling)",
+		"[С нами ты всегда найдёшь причину улыбнуться!](t.me/nomoredoomscrolling)",
+		"[Подними себе настроение на нашем канале!](t.me/nomoredoomscrolling)",
+		"[Позитивные истории ждут тебя! Присоединяйся!](t.me/nomoredoomscrolling)",
+	}
+
+	rand.New(rand.NewSource(time.Now().UnixNano()))
+	randomIndex := rand.Intn(len(messageEndings))
+
+	return messageEndings[randomIndex]
+}
